Refactor of App.js file
