+++ src/App.js
@@ -2,152 +2,38 @@ import './App.css';
 
 import Box from '@mui/material/Box';
 import React from 'react';
-import Divider from '@mui/material/Divider';
-import ListItem from '@mui/material/ListItem';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
-import List from '@mui/material/List';
-import MuiAppBar from '@mui/material/AppBar';
-import { styled } from '@mui/material/styles';
-import Toolbar from '@mui/material/Toolbar';
-import Typography from '@mui/material/Typography';
-import Drawer from '@mui/material/Drawer';
-import IconButton from '@mui/material/IconButton';
-import MenuIcon from '@mui/icons-material/Menu';
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import AccountBoxIcon from '@mui/icons-material/AccountBox';
-import DirectionsRunIcon from '@mui/icons-material/DirectionsRun';
-import EventIcon from '@mui/icons-material/Event';
-import AnalyticsIcon from '@mui/icons-material/Analytics';
 
-import TrainingList from './Components/TrainingList';
-import CustomerList from './Components/Customerlist';
-import Calendar from './Components/Calendar';
-import Statistics from './Components/Statistics';
+import CustomAppBar from './Components/AppbarComponents/CustomAppBar';
+import CustomDrawer from './Components/AppbarComponents/CustomDrawer';
+import MainContent from './Components/AppbarComponents/MainContent';
 
-const drawerWidth = 240;
 
-const Main = styled('main', { shouldForwardProp: (prop) => prop !== 'open' })(
-  ({ theme, open }) => ({
-    flexGrow: 1,
-    padding: theme.spacing(3),
-    transition: theme.transitions.create('margin', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-    marginLeft: `-${drawerWidth}px`,
-    ...(open && {
-      transition: theme.transitions.create('margin', {
-        easing: theme.transitions.easing.easeOut,
-        duration: theme.transitions.duration.enteringScreen,
-      }),
-      marginLeft: 0,
-    }),
-  }),
-);
 
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop) => prop !== 'open',
-})(({ theme, open }) => ({
-  transition: theme.transitions.create(['margin', 'width'], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    width: `calc(100% - ${drawerWidth}px)`,
-    marginLeft: `${drawerWidth}px`,
-    transition: theme.transitions.create(['margin', 'width'], {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}));
 
-const DrawerHeader = styled('div')(({ theme }) => ({
-  display: 'flex',
-  alignItems: 'center',
-  padding: theme.spacing(0, 1),
-  // necessary for content to be below app bar
-  ...theme.mixins.toolbar,
-  justifyContent: 'flex-end',
-}));
 
 function App() {
-  const [value, setValue] = React.useState('Customers');
+  const [selectedCategory, setSelectedCategory] = React.useState('Customers');
 
-  const [open, setOpen] = React.useState(false);
+  const [isDrawerOpen, setIsDrawerOpen] = React.useState(false);
 
   const handleDrawerOpen = () => {
-      setOpen(true);
-  };
-    
-  const handleDrawerClose = () => {
-      setOpen(false);
+    setIsDrawerOpen(true);
   };
 
-  const handleMy = (text) => {
-    setValue(text);
+  const handleDrawerClose = () => {
+    setIsDrawerOpen(false);
+  };
+
+  const changeSelectedCategory = (text) => {
+    setSelectedCategory(text);
   };
 
   return (
     <div className="App">
       <Box sx={{ display: 'flex' }}>
-        <AppBar position="fixed">
-          <Toolbar>
-            <IconButton
-              color="inherit"
-              aria-label="open drawer"
-              onClick={handleDrawerOpen}
-              edge="start"
-              sx={{ mr: 2, ...(open && { display: 'none' }) }}
-            >            
-              <MenuIcon />
-            </IconButton>
-            <Typography variant="h6" component="div">
-              Personal Trainer
-            </Typography>
-          </Toolbar>
-        </AppBar>
-        <Drawer
-          sx={{
-            width: drawerWidth,
-            flexShrink: 0,
-            '& .MuiDrawer-paper': {
-              width: drawerWidth,
-              boxSizing: 'border-box',
-            },
-          }}
-          variant="persistent"
-          anchor="left"
-          open={open}
-        >
-          <DrawerHeader>
-            <IconButton onClick={handleDrawerClose}>
-              {<ChevronLeftIcon />}
-            </IconButton>
-          </DrawerHeader>
-          <Divider />
-          <List>
-            {['Customers', 'Trainings', 'Calendar', 'Statistics'].map((text, index) => (
-              <ListItem  button onClick={()=>handleMy(text)}>
-                <ListItemIcon>
-                  {index === 0 && <AccountBoxIcon />}
-                  {index === 1 && <DirectionsRunIcon />}
-                  {index === 2 && <EventIcon />}
-                  {index === 3 && <AnalyticsIcon />}
-                </ListItemIcon>
-                <ListItemText primary={text} />
-              </ListItem>
-            ))}
-          </List>
-        </Drawer>
-        <Main open={open}>
-          <DrawerHeader />
-          {value === 'Customers' && <CustomerList />}
-          {value === 'Trainings' && <TrainingList />}
-          {value === 'Calendar' && <Calendar />}
-          {value === 'Statistics' && <Statistics />}
-        </Main>
+        <CustomAppBar handleDrawerOpen={handleDrawerOpen} isDrawerOpen={isDrawerOpen} />
+        <CustomDrawer changeSelectedCategory={changeSelectedCategory} handleDrawerClose={handleDrawerClose} isDrawerOpen={isDrawerOpen} />
+        <MainContent isDrawerOpen={isDrawerOpen} selectedCategory={selectedCategory} />
       </Box>
     </div>
   );
